Cleaned up dependencies. ...
Some whitespace, and added version constraints to the `bytestring`
dependency. The constraints are probably pretty conservative because we
only use the `Data.ByteString.Lazy.ByteString` symbol, but also probably
plenty liberal because stack LTS 2.22, (which is the oldest LTS I've
seen used), specifies bytestring-0.10.4.0, which is even newer than the
oldest version we are allowing.