Use Template Haskell based logging functions. ...
I had been trying to avoid introducing template haskell into this
library, however there are a couple of really good reasons why we
should use the template haskell functions.

The first is that without template haskell, the module and file location
show up as `<unknown>` in the log files. While this information may
not turn out to be super useful to programmer trying to debug a
program, it is extremely frustrating to see `<unknown>` and not know
whether the location information would have been helpful or not.
Providing the location lessens the user's uncertainty about what they
are looking at.

Secondly, without the module information, it is impossible to
configure a logging system to take special action for log messages
originating from that module, e.g. like using a higher or lower
logging level than the rest of the system.